Clearly separate old and new mismi interfaces
Also remove unused bits from legacy `Aws.*` modules
