+++ mismi-s3/src/Mismi/S3.hs
@@ -1,8 +1,8 @@
+-- | Import this module to use new `amazonka`-based Mismi
 module Mismi.S3 (
     module X
   ) where
 
-import           Mismi.S3.Aws.Control as X
-import           Mismi.S3.Aws.Data as X
-import           Mismi.S3.Aws.Default as X
+import           Mismi.Control.Amazonka as X
 import           Mismi.S3.Data as X
+import           Mismi.S3.Default as X
