+++ mismi-s3/test/Test/IO/Mismi/S3/Control.hs
@@ -6,7 +6,7 @@ module Test.IO.Mismi.S3.Control where
 
 import           Data.Text
 
-import           Mismi.S3
+import           Mismi.S3.Aws
 
 import           P
