+++ mismi-s3/test/Test/Reliability/Mismi/S3/Aws/Commands.hs
@@ -3,7 +3,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -fno-warn-missing-signatures #-}
-module Test.Reliability.Mismi.S3.Amazonka where
+module Test.Reliability.Mismi.S3.Aws.Commands where
 
 import           Control.Monad.Catch
 
@@ -12,9 +12,7 @@ import qualified Data.Text as T
 
 import           Disorder.Corpus
 
-import           Mismi.S3.Data
-import qualified Mismi.S3.Aws.Default as S3
-import           Mismi.S3.Default
+import           Mismi.S3.Aws
 
 import           P
 
@@ -25,19 +23,18 @@ import           Test.Reliability.Reliability
 import           Test.QuickCheck
 import           Test.QuickCheck.Instances ()
 
-prop_sync = forAll (elements muppets) $ \m -> testAWS' $ \a b i -> do
+prop_list :: Property
+prop_list = forAll (elements muppets) $ \m -> testS3 $ \a i -> do
   createFiles a m i
-  syncWithMode OverwriteSync a b 10
-  mapM_ (\e -> exists e >>= \e' -> when (e' == False) (throwM $ userError "Output files do not exist")) (files a m i)
+  replicateM_ 100 (list a >>= \z -> when (length z /=  i) (throwM $ userError "List is not the same as original response"))
   pure $ True === True
 
-createFiles :: Address -> Text -> Int -> AWS ()
+createFiles :: Address -> Text -> Int -> S3Action ()
 createFiles prefix name n = do
-  mapM_ (flip S3.write "data") $ files prefix name n
-
-files :: Address -> Text -> Int -> [Address]
-files prefix name n =
-  fmap (\i -> withKey (</> Key (name <> "-" <> (T.pack $ show i))) prefix) [1..n]
+  mapM_ (\i ->
+          flip write "data" $
+            withKey (</> Key (name <> "-" <> (T.pack $ show i))) prefix
+        ) [1..n]
 
 return []
 tests :: IO Bool
