+++ mismi-s3/test/Test/Reliability/Mismi/S3/Commands.hs
@@ -23,18 +23,19 @@ import           Test.Reliability.Reliability
 import           Test.QuickCheck
 import           Test.QuickCheck.Instances ()
 
-prop_list :: Property
-prop_list = forAll (elements muppets) $ \m -> testS3 $ \a i -> do
+prop_sync = forAll (elements muppets) $ \m -> testAWS' $ \a b i -> do
   createFiles a m i
-  replicateM_ 100 (list a >>= \z -> when (length z /=  i) (throwM $ userError "List is not the same as original response"))
+  syncWithMode OverwriteSync a b 10
+  mapM_ (\e -> exists e >>= \e' -> when (e' == False) (throwM $ userError "Output files do not exist")) (files a m i)
   pure $ True === True
 
-createFiles :: Address -> Text -> Int -> S3Action ()
+createFiles :: Address -> Text -> Int -> AWS ()
 createFiles prefix name n = do
-  mapM_ (\i ->
-          flip write "data" $
-            withKey (</> Key (name <> "-" <> (T.pack $ show i))) prefix
-        ) [1..n]
+  mapM_ (flip write "data") $ files prefix name n
+
+files :: Address -> Text -> Int -> [Address]
+files prefix name n =
+  fmap (\i -> withKey (</> Key (name <> "-" <> (T.pack $ show i))) prefix) [1..n]
 
 return []
 tests :: IO Bool
