+++ mismi-s3/test/test-io.hs
@@ -1,16 +1,16 @@
 import           Disorder.Core.Main
 
+import qualified Test.IO.Mismi.S3.Aws.Commands
 import qualified Test.IO.Mismi.S3.Commands
 import qualified Test.IO.Mismi.S3.Control
-import qualified Test.IO.Mismi.S3.Amazonka
 import qualified Test.IO.Mismi.S3.Internal
 
 
 main :: IO ()
 main =
   disorderMain [
-      Test.IO.Mismi.S3.Commands.tests
+      Test.IO.Mismi.S3.Aws.Commands.tests
+    , Test.IO.Mismi.S3.Commands.tests
     , Test.IO.Mismi.S3.Control.tests
-    , Test.IO.Mismi.S3.Amazonka.tests
     , Test.IO.Mismi.S3.Internal.tests
     ]
