+++ mismi-s3/test/test-reliability.hs
@@ -1,12 +1,12 @@
 import           Disorder.Core.Main
 
+import qualified Test.Reliability.Mismi.S3.Aws.Commands
 import qualified Test.Reliability.Mismi.S3.Commands
-import qualified Test.Reliability.Mismi.S3.Amazonka
 
 
 main :: IO ()
 main =
   disorderMain [
-      Test.Reliability.Mismi.S3.Commands.tests
-    , Test.Reliability.Mismi.S3.Amazonka.tests
+      Test.Reliability.Mismi.S3.Aws.Commands.tests
+    , Test.Reliability.Mismi.S3.Commands.tests
     ]
